# Fix max nan