* Adds nan check to Tensor.max()
* Calls to max can pass in "ignore_nan=True/False" argument
* Added max nan CI tests

GPU=1 python3 test/test_speed_v_torch.py TestSpeed.test_max on 3090

"ignore_nan=False"
**This Patch:**
`
max   4096x 4096 3.79 ms (4.43 GFLOPS 17.71 GB/s) in torch, 9.52 ms (1.76 GFLOPS  7.05 GB/s) in tinygrad, 2.51x slower      16.78 MOPS 67.11 MB`
.
**Master:**
`max   4096x 4096    3.72 ms (4.51 GFLOPS 18.04 GB/s) in torch, 9.52 ms ( 1.76 GFLOPS  7.05 GB/s) in tinygrad,  2.56x slower  16.78 MOPS    67.11 MB`

**Torch anomoly detection** : [CLASStorch.autograd.set_detect_anomaly(mode, check_nan=True)](https://pytorch.org/docs/stable/autograd.html#torch.autograd.detect_anomaly)