@tinyb0t 2023-07-20
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/tensor.py                         3               1
------------------------------------------------------------
lines added in the tinygrad folder: 2
```