@wozeparrot 2023-07-24
So this breaks the backwards pass when `ignore_nan=False` due to the use of `!=` in `Tensor.isnan` which uses `==` in it's implementation which doesn't have a defined backwards pass.